update handle simple no network error & move common dir
