+++ app/src/main/java/com/example/githubuserbrowser/feature/detail/DetailScreen.kt
@@ -1,5 +1,6 @@
 package com.example.githubuserbrowser.feature.detail
 
+import android.widget.Toast
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -22,19 +23,27 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.githubuserbrowser.feature.detail.components.AvatarImageView
-import com.example.githubuserbrowser.feature.detail.components.HyperlinkText
-import com.example.githubuserbrowser.feature.detail.components.IndeterminateCircularIndicator
+import com.example.githubuserbrowser.core.designsystem.components.AvatarImageView
+import com.example.githubuserbrowser.core.designsystem.components.HyperlinkText
+import com.example.githubuserbrowser.core.designsystem.components.IndeterminateCircularIndicator
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun UserDetailScreen(uiState: UIState, onBackPressed: () -> Unit) {
+    val context = LocalContext.current
+    LaunchedEffect(uiState.errorText, uiState.userDetail?.userName) {
+        if(uiState.errorText != null) {
+            Toast.makeText(context, uiState.errorText.toString(), Toast.LENGTH_SHORT).show()
+        }
+    }
     Scaffold(
         topBar = {
             CenterAlignedTopAppBar(
@@ -66,7 +75,12 @@ fun UserDetailScreen(uiState: UIState, onBackPressed: () -> Unit) {
                 .padding(4.dp, contentPadding.calculateTopPadding(), 0.dp, contentPadding.calculateBottomPadding())
         ) {
             uiState.userDetail?.avatarUrl?.let {
-                AvatarImageView(Modifier.height(250.dp).width(250.dp).clip(CircleShape).align(Alignment.CenterHorizontally), it)
+                AvatarImageView(
+                    Modifier
+                        .height(250.dp)
+                        .width(250.dp)
+                        .clip(CircleShape)
+                        .align(Alignment.CenterHorizontally), it)
             }
             if (uiState.isLoading) {
                 Box(contentAlignment = Alignment.Center, modifier = Modifier.fillMaxSize()) {
@@ -77,46 +91,47 @@ fun UserDetailScreen(uiState: UIState, onBackPressed: () -> Unit) {
                             .padding(8.dp)
                     )
                 }
-            } else
-                 Spacer(modifier = Modifier.height(16.dp))
-                 uiState.userDetail?.let {
-                     Spacer(modifier = Modifier.height(8.dp))
-                     Text(
-                         text = it.userName,
-                         style = MaterialTheme.typography.bodyMedium,
-                         fontWeight = FontWeight.Bold,
-                         fontSize = 24.sp,
-                         modifier = Modifier.align(Alignment.CenterHorizontally)
-                     )
-                     Spacer(modifier = Modifier.height(16.dp))
+            } else {
+                Spacer(modifier = Modifier.height(16.dp))
+                uiState.userDetail?.let {
+                    Spacer(modifier = Modifier.height(8.dp))
+                    Text(
+                        text = it.userName,
+                        style = MaterialTheme.typography.bodyMedium,
+                        fontWeight = FontWeight.Bold,
+                        fontSize = 24.sp,
+                        modifier = Modifier.align(Alignment.CenterHorizontally)
+                    )
+                    Spacer(modifier = Modifier.height(16.dp))
 
-                     if(it.location?.isNotEmpty() == true) {
-                         Text(
-                             text = "Location: " + it.location,
-                             style = MaterialTheme.typography.bodyMedium,
-                             fontSize = 16.sp
-                         )
-                     }
-                     Spacer(modifier = Modifier.height(8.dp))
-                     Text(
-                         text = "Followers " + it.followers,
-                         style = MaterialTheme.typography.bodyMedium,
-                         fontSize = 16.sp
-                     )
-                     Spacer(modifier = Modifier.height(8.dp))
-                     Text(
-                         text = "Following " + it.following,
-                         style = MaterialTheme.typography.bodyMedium,
-                         fontSize = 16.sp
-                     )
-                     Spacer(modifier = Modifier.height(8.dp))
+                    if (it.location?.isNotEmpty() == true) {
+                        Text(
+                            text = "Location: " + it.location,
+                            style = MaterialTheme.typography.bodyMedium,
+                            fontSize = 16.sp
+                        )
+                    }
+                    Spacer(modifier = Modifier.height(8.dp))
+                    Text(
+                        text = "Followers " + it.followers,
+                        style = MaterialTheme.typography.bodyMedium,
+                        fontSize = 16.sp
+                    )
+                    Spacer(modifier = Modifier.height(8.dp))
+                    Text(
+                        text = "Following " + it.following,
+                        style = MaterialTheme.typography.bodyMedium,
+                        fontSize = 16.sp
+                    )
+                    Spacer(modifier = Modifier.height(8.dp))
 
-                     it.landingPageUrl.let { homepageUrl ->
-                         val showText = "For more information: $homepageUrl"
-                         HyperlinkText(fullText = showText, linkText = listOf(homepageUrl), hyperlinks = listOf(homepageUrl), fontSize = 16.sp)
-                     }
-                     Spacer(modifier = Modifier.height(8.dp))
-                 }
+                    it.landingPageUrl.let { homepageUrl ->
+                        val showText = "For more information: $homepageUrl"
+                        HyperlinkText(fullText = showText, linkText = listOf(homepageUrl), hyperlinks = listOf(homepageUrl), fontSize = 16.sp)
+                    }
+                    Spacer(modifier = Modifier.height(8.dp))
+                }
+            }
         }
     }
 }
