+++ app/src/main/java/com/example/githubuserbrowser/feature/detail/DetailViewModel.kt
@@ -2,12 +2,9 @@ package com.example.githubuserbrowser.feature.detail
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import androidx.paging.PagingData
-import androidx.paging.cachedIn
 import com.example.githubuserbrowser.core.data.model.UserModel
 import com.example.githubuserbrowser.core.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.catch
@@ -28,15 +25,21 @@ class DetailViewModel @Inject constructor(
         userRepository.getUserDetail(userName).onStart {
             _uiState.value = _uiState.value.copy(isLoading = true)
         }.catch {
-            _uiState.value = _uiState.value.copy(isLoading = false)
+            _uiState.value = _uiState.value.copy(isLoading = false, errorText = "Network is unavailable")
         }.onEach { res ->
             _uiState.value = _uiState.value.copy(isLoading = false, userDetail = res)
         }.launchIn(viewModelScope)
     }
+
+    fun clearUiState() {
+        _uiState.value = UIState()
+    }
+
 }
 
 data class UIState(
     var isLoading: Boolean = false,
-    var userDetail: UserModel? = null
+    var userDetail: UserModel? = null,
+    var errorText: String? = null
 )
