+++ app/src/main/java/com/example/githubuserbrowser/feature/detail/navigation/UserDetailDestination.kt
@@ -27,6 +27,7 @@ fun NavGraphBuilder.userDetailScreen(
         UserDetailScreen(
             uiState,
             onBackPressed = {
+                viewModel.clearUiState()
                 onBackPressed()
             }
         )
