+++ app/src/main/java/com/example/githubuserbrowser/feature/home/HomeScreen.kt
@@ -2,12 +2,8 @@ package com.example.githubuserbrowser.feature.home
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.rememberLazyGridState
@@ -20,7 +16,7 @@ import androidx.paging.LoadState
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.itemKey
 import com.example.githubuserbrowser.core.data.model.UserModel
-import com.example.githubuserbrowser.feature.home.component.UserCard
+import com.example.githubuserbrowser.core.designsystem.components.UserCard
 
 @Composable
 fun HomeScreen(
